Conflicts:
	src/sql/drivers/sqlite/qsql_sqlite.cpp

+++ src/opengl/gl2paintengineex/qglengineshadersource_p.h
@@ -238,7 +238,7 @@ static const char* const qglslPositionWithRadialGradientBrushVertexShader = "\n\
     uniform   mediump vec2      halfViewportSize; \n\
     uniform   highp   mat3      brushTransform; \n\
     uniform   highp   vec2      fmp; \n\
-    uniform   highp   vec3      bradius; \n\
+    uniform   mediump vec3      bradius; \n\
     varying   highp   float     b; \n\
     varying   highp   vec2      A; \n\
     void setPosition(void) \n\
@@ -264,7 +264,7 @@ static const char* const qglslRadialGradientBrushSrcFragmentShader = "\n\
     uniform   highp   float     sqrfr; \n\
     varying   highp   float     b; \n\
     varying   highp   vec2      A; \n\
-    uniform   highp   vec3      bradius; \n\
+    uniform   mediump vec3      bradius; \n\
     lowp vec4 srcPixel() \n\
     { \n\
         highp float c = sqrfr-dot(A, A); \n\
